Now everything will be minified
