+++ public/css/appearance.css
@@ -1,4 +1,1 @@
-#manifesto .scroll-appearance {
-  transform: translateY(100%);
-  opacity: 0;
-}
+#manifesto .scroll-appearance{transform:translateY(100%);opacity:0}
