+++ public/css/fonts.css
@@ -1,6 +1,1 @@
-@font-face {
-  font-family: "Brandon";
-  src: url("../fonts/Brandon.eot");
-  src: url("../fonts/Brandon.eot?#iefix") format("embedded-opentype"), url("../fonts/Brandon.woff") format("woff"), url("../fonts/Brandon.ttf") format("truetype");
-  font-style: bold;
-}
+@font-face{font-family:"Brandon";src:url("../fonts/Brandon.eot");src:url("../fonts/Brandon.eot?#iefix") format("embedded-opentype"),url("../fonts/Brandon.woff") format("woff"),url("../fonts/Brandon.ttf") format("truetype");font-style:bold}
