+++ public/css/footer.css
@@ -1,150 +1,1 @@
-#footer {
-  position: relative;
-  width: 100%;
-  height: 100%;
-  min-width: minimum-width;
-}
-#footer .wrapper {
-  position: absolute;
-  top: 45%;
-  transform: translate3d(0, -50%, 0);
-  display: block;
-  margin: auto;
-  width: 100%;
-  transition: padding-transition, opacity-transition;
-}
-#footer .wrapper h2 {
-  font: 30px Brandon;
-  text-transform: uppercase;
-  text-align: center;
-  color: text-color;
-}
-#footer .cross {
-  width: 40px;
-  height: 40px;
-  transform: rotateZ(45deg);
-  margin: 10px auto 0 auto;
-}
-#footer .cross:before {
-  content: '';
-  position: absolute;
-  height: 1px;
-  top: 50%;
-  right: 0%;
-  left: 0%;
-  background-color: text-color;
-}
-#footer .cross:after {
-  content: '';
-  position: absolute;
-  width: 1px;
-  top: 0%;
-  bottom: 0%;
-  left: 50%;
-  background-color: text-color;
-}
-#footer #social {
-  position: absolute;
-  display: table;
-  margin-top: 26px;
-  min-width: 600px;
-  left: 50%;
-  transform: translateX(-50%);
-}
-#footer #social div {
-  display: table-cell;
-}
-#footer #social div a {
-  font: 15px Inconsolata;
-  margin-right: 25px;
-  margin-left: 25px;
-  position: relative;
-}
-#footer #social div figure {
-  position: absolute;
-  width: 17px;
-  height: 17px;
-  background-size: contain;
-  background-repeat: no-repeat;
-  margin-top: 1px;
-}
-#footer #social #dribbble figure {
-  background-image: url("dribbble-icon");
-}
-#footer #social #github figure {
-  background-image: url("github-icon");
-}
-#footer #social #twitter figure {
-  background-image: url("twitter-icon");
-}
-#footer #social #resume figure {
-  background-image: url("monogram-icon");
-}
-#footer #next-project {
-  position: absolute;
-  right: 30px;
-  bottom: 25px;
-  font: header-font;
-  color: #fff;
-  padding-right: 50px;
-  padding-bottom: 7.5px;
-}
-#footer #next-project:after {
-  display: none;
-}
-#footer #next-project:before {
-  content: '';
-  position: absolute;
-  background-color: text-color;
-  height: glitch-size;
-  bottom: 0%;
-  right: 0%;
-  left: 35%;
-  transition: left glitch-transition;
-}
-#footer #next-project:hover:before {
-  left: 0%;
-}
-@media (max-width: minimum-width) {
-  #footer {
-    width: 100vw;
-    height: 100vh;
-    min-width: 100vw;
-    min-height: 690px;
-  }
-  #footer .wrapper {
-    width: 70%;
-    top: 50%;
-    left: 15%;
-  }
-  #footer #social {
-    position: relative;
-    display: inline;
-    width: 100%;
-    min-width: 100%;
-    text-align: center;
-    top: 23px;
-    left: auto;
-    right: auto;
-  }
-  #footer #social div {
-    display: block;
-    margin-bottom: 20px;
-    margin-left: 10px;
-  }
-  #footer #social div a {
-    font: 15px Inconsolata;
-    margin-right: 25px;
-    margin-left: 25px;
-    position: relative;
-  }
-  #footer #social div figure {
-    display: inline;
-    position: absolute;
-    width: 17px;
-    height: 17px;
-    background-size: contain;
-    background-repeat: no-repeat;
-    margin-top: 1px;
-  }
-}
+#footer{position:relative;width:100%;height:100%;min-width:minimum-width}#footer .wrapper{position:absolute;top:45%;transform:translate3d(0,-50%,0);display:block;margin:auto;width:100%;transition:padding-transition,opacity-transition}#footer .wrapper h2{font:30px Brandon;text-transform:uppercase;text-align:center;color:text-color}#footer .cross{width:40px;height:40px;transform:rotateZ(45deg);margin:10px auto 0 auto}#footer .cross:before{content:'';position:absolute;height:1px;top:50%;right:0;left:0;background-color:text-color}#footer .cross:after{content:'';position:absolute;width:1px;top:0;bottom:0;left:50%;background-color:text-color}#footer #social{position:absolute;display:table;margin-top:26px;min-width:600px;left:50%;transform:translateX(-50%)}#footer #social div{display:table-cell}#footer #social div a{font:15px Inconsolata;margin-right:25px;margin-left:25px;position:relative}#footer #social div figure{position:absolute;width:17px;height:17px;background-size:contain;background-repeat:no-repeat;margin-top:1px}#footer #social #dribbble figure{background-image:url("dribbble-icon")}#footer #social #github figure{background-image:url("github-icon")}#footer #social #twitter figure{background-image:url("twitter-icon")}#footer #social #resume figure{background-image:url("monogram-icon")}#footer #next-project{position:absolute;right:30px;bottom:25px;font:header-font;color:#fff;padding-right:50px;padding-bottom:7.5px}#footer #next-project:after{display:none}#footer #next-project:before{content:'';position:absolute;background-color:text-color;height:glitch-size;bottom:0;right:0;left:35%;transition:left glitch-transition}#footer #next-project:hover:before{left:0}@media(max-width:minimum-width){#footer{width:100vw;height:100vh;min-width:100vw;min-height:690px}#footer .wrapper{width:70%;top:50%;left:15%}#footer #social{position:relative;display:inline;width:100%;min-width:100%;text-align:center;top:23px;left:auto;right:auto}#footer #social div{display:block;margin-bottom:20px;margin-left:10px}#footer #social div a{font:15px Inconsolata;margin-right:25px;margin-left:25px;position:relative}#footer #social div figure{display:inline;position:absolute;width:17px;height:17px;background-size:contain;background-repeat:no-repeat;margin-top:1px}}
