+++ public/css/glitches.css
@@ -1,124 +1,1 @@
-.corner-glitch {
-  position: absolute;
-  width: 50px;
-  height: 30px;
-}
-.corner-glitch:before {
-  content: '';
-  position: absolute;
-  background-color: text-color;
-  width: 80%;
-  height: glitch-size;
-  top: 0%;
-  left: 0%;
-}
-.corner-glitch:after {
-  content: '';
-  position: absolute;
-  background-color: text-color;
-  width: glitch-size;
-  height: 90%;
-  right: 0%;
-  bottom: 0%;
-}
-.top-corner-glitch {
-  top: 37.5px;
-  right: corner-glitch-side;
-}
-.bottom-corner-glitch {
-  left: corner-glitch-side;
-  bottom: corner-glitch-up;
-  transform: rotate(180deg);
-}
-.disclosure {
-  position: absolute;
-  opacity: disabled-text-opacity;
-  width: 90px;
-}
-.disclosure p {
-  font: glitch-font;
-  color: text-color;
-}
-.disclosure:after {
-  content: '';
-  position: absolute;
-  background-color: text-color;
-  width: 85%;
-  height: glitch-size;
-  bottom: 0%;
-  right: 0%;
-}
-.disclosure-top {
-  top: 17.5px;
-  right: 325px;
-}
-.disclosure-disclaimer {
-  left: corner-glitch-side;
-  bottom: 80px;
-  width: 105px;
-}
-.horizontal-glitch {
-  position: absolute;
-  width: 35px;
-  height: 100px;
-}
-.horizontal-glitch:before {
-  content: '';
-  position: absolute;
-  width: glitch-size;
-  height: 25%;
-  right: 0%;
-  top: 0%;
-  background-color: text-color;
-  opacity: disabled-text-opacity;
-}
-.horizontal-glitch:after {
-  content: '';
-  position: absolute;
-  width: glitch-size;
-  height: 50%;
-  right: 0%;
-  top: 50%;
-  background-color: text-color;
-  opacity: disabled-text-opacity;
-}
-.horizontal-glitch div {
-  content: '';
-  position: absolute;
-  height: glitch-size;
-  top: 40%;
-  right: 0%;
-  left: 0%;
-  background-color: text-color;
-  opacity: disabled-text-opacity;
-}
-.story-horizontal-glitch {
-  top: 220px;
-  right: 100px;
-}
-#dot-separator {
-  position: absolute;
-  width: 96.48vw;
-  height: 11px;
-  bottom: corner-glitch-up -10px;
-  left: corner-glitch-side;
-  min-width: minimum-width;
-  opacity: 0.5;
-}
-#dot-separator div {
-  float: left;
-  width: glitch-size;
-  height: glitch-size;
-  margin-right: 10px;
-  margin-bottom: 10px;
-  display: inline-block;
-  background-color: text-color;
-}
-@media (max-width: minimum-width) {
-  .disclosure {
-    display: none;
-  }
-  #dot-separator {
-    display: none;
-  }
-}
+.corner-glitch{position:absolute;width:50px;height:30px}.corner-glitch:before{content:'';position:absolute;background-color:text-color;width:80%;height:glitch-size;top:0;left:0}.corner-glitch:after{content:'';position:absolute;background-color:text-color;width:glitch-size;height:90%;right:0;bottom:0}.top-corner-glitch{top:37.5px;right:corner-glitch-side}.bottom-corner-glitch{left:corner-glitch-side;bottom:corner-glitch-up;transform:rotate(180deg)}.disclosure{position:absolute;opacity:disabled-text-opacity;width:90px}.disclosure p{font:glitch-font;color:text-color}.disclosure:after{content:'';position:absolute;background-color:text-color;width:85%;height:glitch-size;bottom:0;right:0}.disclosure-top{top:17.5px;right:325px}.disclosure-disclaimer{left:corner-glitch-side;bottom:80px;width:105px}.horizontal-glitch{position:absolute;width:35px;height:100px}.horizontal-glitch:before{content:'';position:absolute;width:glitch-size;height:25%;right:0;top:0;background-color:text-color;opacity:disabled-text-opacity}.horizontal-glitch:after{content:'';position:absolute;width:glitch-size;height:50%;right:0;top:50%;background-color:text-color;opacity:disabled-text-opacity}.horizontal-glitch div{content:'';position:absolute;height:glitch-size;top:40%;right:0;left:0;background-color:text-color;opacity:disabled-text-opacity}.story-horizontal-glitch{top:220px;right:100px}#dot-separator{position:absolute;width:96.48vw;height:11px;bottom:corner-glitch-up -10px;left:corner-glitch-side;min-width:minimum-width;opacity:.5}#dot-separator div{float:left;width:glitch-size;height:glitch-size;margin-right:10px;margin-bottom:10px;display:inline-block;background-color:text-color}@media(max-width:minimum-width){.disclosure{display:none}#dot-separator{display:none}}
