+++ public/css/header.css
@@ -1,93 +1,1 @@
-.detail header,
-#manifesto header,
-#myself header {
-  position: fixed;
-  width: 100%;
-  height: 60px;
-  min-width: minimum-width;
-  top: 0;
-  left: 0;
-  margin: 0 auto;
-  z-index: 1;
-  background-color: header-background-color;
-  transition: initial-opacity-transition;
-}
-.detail .email-link-wrapper,
-#manifesto .email-link-wrapper,
-#myself .email-link-wrapper {
-  position: absolute;
-  margin-top: 32.5px;
-  transform: translateY(-50%);
-  right: 105px;
-  cursor: pointer;
-}
-.detail .email,
-#manifesto .email,
-#myself .email {
-  font: header-font;
-  cursor: pointer;
-}
-.detail #header-glitches,
-#manifesto #header-glitches,
-#myself #header-glitches {
-  transition: initial-opacity-transition;
-}
-.detail #back-button,
-#manifesto #back-button,
-#myself #back-button {
-  position: absolute;
-  top: 17.5px;
-  font: header-font;
-  color: #fff;
-  padding-left: 50px;
-  padding-bottom: 7.5px;
-  left: corner-glitch-side;
-  cursor: pointer;
-}
-.detail #back-button:after,
-#manifesto #back-button:after,
-#myself #back-button:after {
-  display: none;
-}
-.detail #back-button:before,
-#manifesto #back-button:before,
-#myself #back-button:before {
-  content: '';
-  position: absolute;
-  background-color: text-color;
-  height: glitch-size;
-  bottom: 0%;
-  left: 0%;
-  right: 35%;
-  transition: right glitch-transition;
-}
-.detail #back-button:hover:before,
-#manifesto #back-button:hover:before,
-#myself #back-button:hover:before {
-  right: 0%;
-}
-#myself #back-button {
-  display: none;
-}
-@media (max-width: minimum-width) {
-  .detail header,
-  #manifesto header,
-  #myself header {
-    width: 100vw;
-    min-width: 100px;
-    max-width: 100vw;
-  }
-  .detail .email-link-wrapper,
-  #manifesto .email-link-wrapper,
-  #myself .email-link-wrapper {
-    position: absolute;
-    margin-top: 32.5px;
-    transform: translate(50%, -50%);
-    right: 50%;
-    white-space: nowrap;
-  }
-  .detail .email-link-wrapper,
-  #manifesto .email-link-wrapper {
-    display: none;
-  }
-}
+.detail header,#manifesto header,#myself header{position:fixed;width:100%;height:60px;min-width:minimum-width;top:0;left:0;margin:0 auto;z-index:1;background-color:header-background-color;transition:initial-opacity-transition}.detail .email-link-wrapper,#manifesto .email-link-wrapper,#myself .email-link-wrapper{position:absolute;margin-top:32.5px;transform:translateY(-50%);right:105px;cursor:pointer}.detail .email,#manifesto .email,#myself .email{font:header-font;cursor:pointer}.detail #header-glitches,#manifesto #header-glitches,#myself #header-glitches{transition:initial-opacity-transition}.detail #back-button,#manifesto #back-button,#myself #back-button{position:absolute;top:17.5px;font:header-font;color:#fff;padding-left:50px;padding-bottom:7.5px;left:corner-glitch-side;cursor:pointer}.detail #back-button:after,#manifesto #back-button:after,#myself #back-button:after{display:none}.detail #back-button:before,#manifesto #back-button:before,#myself #back-button:before{content:'';position:absolute;background-color:text-color;height:glitch-size;bottom:0;left:0;right:35%;transition:right glitch-transition}.detail #back-button:hover:before,#manifesto #back-button:hover:before,#myself #back-button:hover:before{right:0}#myself #back-button{display:none}@media(max-width:minimum-width){.detail header,#manifesto header,#myself header{width:100vw;min-width:100px;max-width:100vw}.detail .email-link-wrapper,#manifesto .email-link-wrapper,#myself .email-link-wrapper{position:absolute;margin-top:32.5px;transform:translate(50%,-50%);right:50%;white-space:nowrap}.detail .email-link-wrapper,#manifesto .email-link-wrapper{display:none}}
