+++ public/css/loader.css
@@ -1,45 +1,1 @@
-#loader {
-  position: fixed;
-  top: 0;
-  right: 0;
-  bottom: 0;
-  left: 0;
-  margin: 0 auto;
-  background-color: background-color;
-  z-index: 2;
-}
-#loader .percentage {
-  position: absolute;
-  top: 45%;
-  right: 15px;
-  font: header-font;
-  color: text-color;
-  transform: translateX(200px);
-}
-#loader .bar {
-  position: absolute;
-  height: 1px;
-  top: 50%;
-  right: 100%;
-  left: 0%;
-  transform: translateY(-50%);
-  background-color: text-color;
-}
-#loader .link-wrapper {
-  position: absolute;
-  bottom: 20px;
-  left: 50%;
-  transform: translate(-50%, 500%);
-  text-align: center;
-  opacity: 0.2;
-}
-#loader .link-wrapper .email {
-  left: 50%;
-  transform: translateX(-50%);
-  padding: 0;
-}
-@media (max-width: minimum-width) {
-  #loader .link-wrapper {
-    display: none;
-  }
-}
+#loader{position:fixed;top:0;right:0;bottom:0;left:0;margin:0 auto;background-color:background-color;z-index:2}#loader .percentage{position:absolute;top:45%;right:15px;font:header-font;color:text-color;transform:translateX(200px)}#loader .bar{position:absolute;height:1px;top:50%;right:100%;left:0;transform:translateY(-50%);background-color:text-color}#loader .link-wrapper{position:absolute;bottom:20px;left:50%;transform:translate(-50%,500%);text-align:center;opacity:.2}#loader .link-wrapper .email{left:50%;transform:translateX(-50%);padding:0}@media(max-width:minimum-width){#loader .link-wrapper{display:none}}
