+++ public/css/loading.css
@@ -1,22 +1,1 @@
-#manifesto .loading-hero-image {
-  transform: translateX(150px);
-  opacity: 0;
-}
-#manifesto .loading-hero-line {
-  transform: translateX(50px);
-  opacity: 0;
-}
-#manifesto .loading-header {
-  opacity: 0;
-}
-.detail .loading-title {
-  transform: translateY(-100%) !important;
-  opacity: 0;
-}
-.detail .loading-headers {
-  transform: translateY(300%) !important;
-  opacity: 0;
-}
-.detail .loading-header {
-  opacity: 0;
-}
+#manifesto .loading-hero-image{transform:translateX(150px);opacity:0}#manifesto .loading-hero-line{transform:translateX(50px);opacity:0}#manifesto .loading-header{opacity:0}.detail .loading-title{transform:translateY(-100%) !important;opacity:0}.detail .loading-headers{transform:translateY(300%) !important;opacity:0}.detail .loading-header{opacity:0}
