+++ public/js/pagination.js
@@ -1,288 +1,1 @@
-// JavaScript for the pagination in the myself page.
-// Developers: Me, Myself and I.
-
-// MARK: - Window events
-
-window.addEventListener('load', function() {
-
-  var triangle = document.id('triangle');
-  var indicator = document.id('indicator');
-  var myself = document.id('sections');
-  var selectedWork = document.id('selected-work');
-  var selectedWrapper = document.id('selected-wrapper');
-  var backTop = document.id('back-top-button');
-
-  var titles = slice(document.classes('text'), 3);
-  var navigators = document.classes('navigator');
-
-  var sections = document.tags('section');
-
-  var position = 0;
-  var currentPosition = position;
-  var currentTime = new Date().getTime();
-  var pastTime = 0;
-  var safety = 700;
-  var shouldAnimate = true;
-  var cubicBezier = 'cubic-bezier(0.88, 0.12, 0.36, 0.79)'
-
-  swipe(myself);
-  positionIndicator();
-
-  backTop.addEventListener('click', function() {
-    currentPosition = position;
-    position = 0;
-    scroll(0);
-  });
-
-  window.addEventListener('resize', function() {
-    resize();
-    positionIndicator();
-  });
-
-  window.addEventListener('focus', function() {
-    resize();
-  });
-
-  document.addEventListener('swipeUp', function(event) {
-    scrollUp()
-  });
-
-  document.addEventListener('swipeDown', function(event) {
-    scrollDown()
-  });
-
-  window.addEventListener('DOMMouseScroll', scrollHandler);
-  window.addEventListener('mousewheel', scrollHandler);
-
-  for (var i = 0; i < navigators.length; i++) {
-    var navigator = navigators[i];
-
-    navigator.addEventListener('mouseover', function() {
-      positionIndicator(this);
-    });
-
-    navigator.addEventListener('mouseout', function() {
-      setTimeout(function() {
-        positionIndicator();
-      }, 25);
-    });
-
-    navigator.addEventListener('click', function() {
-      currentTime = new Date().getTime();
-
-      if (currentTime - pastTime > safety + 800) {
-        currentPosition = position;
-        position = navigators.indexOf(this);
-
-        scroll(position * window.innerHeight);
-  		}
-    });
-  }
-
-  // MARK: - Methods for the document scrolling.
-
-  function scrollHandler(event) {
-    currentTime = new Date().getTime();
-
-    if (currentTime - pastTime < safety + 800) {
-      event.preventDefault();
-      return;
-    }
-
-    event.preventDefault();
-    var delta = event.wheelDelta || -event.detail;
-
-    delta > 0 ? scrollUp() : scrollDown();
-  }
-
-  function scrollUp() {
-    if (position <= 0) { return; }
-
-    currentPosition = position;
-    position = position - 1;
-    scroll(position * window.innerHeight);
-  }
-
-  function scrollDown() {
-    if (position >= sections.length - 1) { return; }
-
-    currentPosition = position;
-    position = position + 1;
-    scroll(position * window.innerHeight);
-  }
-
-  function scroll(point, animate) {
-    if (point > window.innerHeight * sections.length - window.innerHeight
-      || point < 0 || currentPosition == position) {
-        return;
-    }
-
-    if (animate == undefined) {
-      scrollingTitles();
-    }
-
-    var remove = position == 0;
-
-    toggle(selectedWork, 'selected-work-scroll', remove);
-    toggle(selectedWrapper, 'selected-wrapper-scroll', remove);
-    toggle(backTop, 'back-top-scroll', remove);
-
-    myself.style.transition = animate === undefined ? 'transform 0.8s ' + cubicBezier : '';
-    myself.style.transform = 'translate3d(0, ' + -point + 'px, 0)';
-
-    getTransform(myself);
-    positionIndicator();
-
-    pastTime = currentTime;
-  }
-
-  function resize() {
-    var windowHeight = '100%';
-    for (var i = 0; i < this.sections.length; i++) {
-      var section = this.sections[i];
-      section.style.height = windowHeight;
-    }
-
-    myself.style.height = windowHeight;
-
-    shouldAnimate = false;
-    scroll(position * window.innerHeight, false);
-  }
-
-  function scrollingTitles() {
-    var initialTitles = titles[currentPosition];
-    var finalTitles = titles[position];
-    var titleTransition = 'top 1s ease';
-    var timeout = 800;
-    var starter = 200;
-    var transitionPadding = 80;
-    var up = currentPosition - position > 0;
-
-    var index = 0;
-    loop();
-
-    function loop() {
-      var initialTitle = up
-      ? initialTitles[initialTitles.length - index - 1] : initialTitles[index];
-      var finalTitle = up
-      ? finalTitles[finalTitles.length - index - 1] : finalTitles[index];
-      var initialClass = up ? 'down' : 'up';
-      var finalClass = up ? 'up' : 'down';
-
-      initialTitle.style.transition = titleTransition;
-      initialTitle.classList.add(initialClass);
-
-      finalTitle.style.transition = '';
-      finalTitle.classList.add(finalClass);
-
-      setTimeout(function() {
-        initialTitle.classList.remove(initialClass);
-      }, timeout);
-
-      setTimeout(function() {
-        finalTitle.style.transition = titleTransition;
-        finalTitle.classList.remove(finalClass);
-      }, starter);
-
-      index = index + 1;
-      setTimeout(function() {
-        if (index < initialTitles.length) { loop(); }
-      }, transitionPadding * index);
-    }
-  }
-
-  function positionIndicator(button) {
-    var navigator = button || navigators[position];
-    var y = navigator.getBoundingClientRect().top;
-    var width = navigator.offsetWidth;
-    var triangleWidth = triangle.offsetWidth;
-
-    indicator.style.transition = shouldAnimate
-    ? 'top 0.6s ' + cubicBezier + ', height 0.6s ' + cubicBezier : '';
-    triangle.style.transition = shouldAnimate ? 'top 0.6s ' + cubicBezier : '';
-
-    shouldAnimate = true;
-
-    setTimeout(function() {
-      indicator.style.top = y + 'px';
-      indicator.style.height = width + 'px';
-
-      if (button === undefined) {
-        triangle.style.top = y + width / 2 - triangleWidth / 2 + 'px';
-      }
-    }, 0);
-
-    indicator.style.transition = 'top 0.6s ' + cubicBezier + ', height 0.6s ' + cubicBezier;
-    triangle.style.transition = 'top 0.6s ' + cubicBezier;
-  }
-
-  // Inspiration from:
-  // github.com/peachananr/purejs-onepage-scroll/blob/master/onepagescroll.js#L173
-
-  function swipe(element) {
-    var initial = 0;
-
-    document.addEventListener('touchstart', function(event) {
-      var touches = event.touches;
-      if (touches && touches.length) {
-        initial = touches[0].pageY;
-        document.addEventListener('touchmove', touchMove);
-      }
-    }, false);
-
-    function touchMove(event) {
-      var touches = event.touches;
-
-      if (touches && touches.length) {
-        event.preventDefault();
-        var delta = initial - touches[0].pageY;
-
-        if (delta >= 50) {
-          var event = new Event('swipeDown');
-          document.dispatchEvent(event);
-        } else if (delta <= -50) {
-          var event = new Event('swipeUp');
-          document.dispatchEvent(event);
-        }
-
-        if (abs(delta) >= 50) {
-          document.removeEventListener('touchmove', touchMove);
-        }
-      }
-    }
-  }
-});
-
-// MARK: - Prevent scrolling
-
-var keys = {
-  32 : 1,
-  37 : 1,
-  38 : 1,
-  39 : 1,
-  40 : 1
-};
-
-function preventDefault(event) {
-  event = event || window.event;
-
-  if (event.preventDefault) {
-    event.preventDefault();
-  }
-
-  event.returnValue = false;
-}
-
-function preventDefaultForScrollKeys(event) {
-  if (keys[event.keyCode]) {
-    preventDefault(event);
-    return false;
-  }
-}
-
-function disableScroll() {
-  window.onmousewheel = preventDefault;
-  document.onkeydown = preventDefaultForScrollKeys;
-}
-
-disableScroll();
+function preventDefault(e){e=e||window.event,e.preventDefault&&e.preventDefault(),e.returnValue=!1}function preventDefaultForScrollKeys(e){if(keys[e.keyCode])return preventDefault(e),!1}function disableScroll(){window.onmousewheel=preventDefault,document.onkeydown=preventDefaultForScrollKeys}window.addEventListener("load",function(){function e(e){if(y=(new Date).getTime(),y-E<L+800)return void e.preventDefault();e.preventDefault();var i=e.wheelDelta||-e.detail;i>0?t():n()}function t(){g<=0||(m=g,g-=1,i(g*window.innerHeight))}function n(){g>=p.length-1||(m=g,g+=1,i(g*window.innerHeight))}function i(e,t){if(!(e>window.innerHeight*p.length-window.innerHeight||e<0||m==g)){void 0==t&&s();var n=0==g;toggle(l,"selected-work-scroll",n),toggle(v,"selected-wrapper-scroll",n),toggle(f,"back-top-scroll",n),u.style.transition=void 0===t?"transform 0.8s "+k:"",u.style.transform="translate3d(0, "+-e+"px, 0)",getTransform(u),r(),E=y}}function o(){for(var e="100%",t=0;t<this.sections.length;t++){var n=this.sections[t];n.style.height=e}u.style.height=e,D=!1,i(g*window.innerHeight,!1)}function s(){function e(){var c=d?t[t.length-a-1]:t[a],u=d?n[n.length-a-1]:n[a],l=d?"down":"up",v=d?"up":"down";c.style.transition=i,c.classList.add(l),u.style.transition="",u.classList.add(v),setTimeout(function(){c.classList.remove(l)},o),setTimeout(function(){u.style.transition=i,u.classList.remove(v)},s),a+=1,setTimeout(function(){a<t.length&&e()},r*a)}var t=w[m],n=w[g],i="top 1s ease",o=800,s=200,r=80,d=m-g>0,a=0;e()}function r(e){var t=e||h[g],n=t.getBoundingClientRect().top,i=t.offsetWidth,o=a.offsetWidth;c.style.transition=D?"top 0.6s "+k+", height 0.6s "+k:"",a.style.transition=D?"top 0.6s "+k:"",D=!0,setTimeout(function(){c.style.top=n+"px",c.style.height=i+"px",void 0===e&&(a.style.top=n+i/2-o/2+"px")},0),c.style.transition="top 0.6s "+k+", height 0.6s "+k,a.style.transition="top 0.6s "+k}function d(e){function t(e){var i=e.touches;if(i&&i.length){e.preventDefault();var o=n-i[0].pageY;if(o>=50){var e=new Event("swipeDown");document.dispatchEvent(e)}else if(o<=-50){var e=new Event("swipeUp");document.dispatchEvent(e)}abs(o)>=50&&document.removeEventListener("touchmove",t)}}var n=0;document.addEventListener("touchstart",function(e){var i=e.touches;i&&i.length&&(n=i[0].pageY,document.addEventListener("touchmove",t))},!1)}var a=document.id("triangle"),c=document.id("indicator"),u=document.id("sections"),l=document.id("selected-work"),v=document.id("selected-wrapper"),f=document.id("back-top-button"),w=slice(document.classes("text"),3),h=document.classes("navigator"),p=document.tags("section"),g=0,m=g,y=(new Date).getTime(),E=0,L=700,D=!0,k="cubic-bezier(0.88, 0.12, 0.36, 0.79)";d(u),r(),f.addEventListener("click",function(){m=g,g=0,i(0)}),window.addEventListener("resize",function(){o(),r()}),window.addEventListener("focus",function(){o()}),document.addEventListener("swipeUp",function(e){t()}),document.addEventListener("swipeDown",function(e){n()}),window.addEventListener("DOMMouseScroll",e),window.addEventListener("mousewheel",e);for(var T=0;T<h.length;T++){var b=h[T];b.addEventListener("mouseover",function(){r(this)}),b.addEventListener("mouseout",function(){setTimeout(function(){r()},25)}),b.addEventListener("click",function(){y=(new Date).getTime(),y-E>L+800&&(m=g,g=h.indexOf(this),i(g*window.innerHeight))})}});var keys={32:1,37:1,38:1,39:1,40:1};disableScroll();
