+++ source/gulpfile.js
@@ -4,11 +4,16 @@ var sourcemaps = require('gulp-sourcemaps');
 var stylus = require('gulp-stylus');
 var mustache = require('gulp-mustache-plus');
 var uglify = require('gulp-uglify');
+var uglifyHTML = require('gulp-htmlmin');
+var uglifyCSS = require('gulp-uglifycss');
 
 gulp.task('stylus', function () {
   gulp.src('./stylus/*.styl')
     .pipe(sourcemaps.init())
     .pipe(stylus())
+    .pipe(uglifyCSS({
+      "uglyComments": true
+    }))
     .pipe(gulp.dest('../public/css'));
 });
 
@@ -21,20 +26,25 @@ gulp.task('mustache', function() {
       detail: './mustache/partials/detail.mustache',
       explanation: './mustache/partials/explanation.mustache',
       footer: './mustache/partials/footer.mustache'
-    })).pipe(gulp.dest('../public'));
+    }))
+    .pipe(uglifyHTML({
+      collapseWhitespace: true
+    }))
+    .pipe(gulp.dest('../public'));
 });
 
-gulp.task('minify', function() {
-  gulp.src('../public/js/*.js')
-  .pipe(uglify())
-  .pipe(gulp.dest('../public/js/'));
+gulp.task('javascript', function() {
+  gulp.src('./js/*.js')
+    .pipe(uglify())
+    .pipe(gulp.dest('../public/js/'));
 });
 
-gulp.task('watch', ['stylus', 'mustache', 'minify'], function() {
-  gulp.watch('./stylus/**/*.styl', ['stylus', 'minify']);
-  gulp.watch('./mustache/**/*.mustache', ['mustache', 'minify']);
-  gulp.watch('./gulpfile.js', ['mustache', 'stylus', 'minify']);
-  gulp.watch('./data/gilabert.json', ['mustache', 'stylus']);
+gulp.task('watch', ['stylus', 'mustache', 'javascript'], function() {
+  gulp.watch('./stylus/**/*.styl', ['stylus']);
+  gulp.watch('./mustache/**/*.mustache', ['mustache']);
+  gulp.watch('./js/*.js', ['javascript']);
+  gulp.watch('./gulpfile.js', ['stylus', 'mustache', 'javascript']);
+  gulp.watch('./data/gilabert.json', ['mustache']);
 });
 
-gulp.task('default', ['stylus', 'mustache', 'minify', 'watch']);
+gulp.task('default', ['stylus', 'mustache', 'javascript', 'watch']);
