+++ source/package.json
@@ -12,6 +12,8 @@
     "gulp-sourcemaps" : "latest",
     "gulp-stylus" : "latest",
     "gulp-mustache-plus" : "latest",
+    "gulp-uglifycss" : "latest",
+    "gulp-htmlmin" : "latest",
     "gulp-uglify" : "latest"
   }
 }
